obs-overlay/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
SQLAlchemy==2.0.40
pydantic==2.11.2
python-multipart==0.0.20
pywebview==5.4
